bq_caapi_ge/pyproject.toml
[project]
name = "bq-caapi-ge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-cloud-geminidataanalytics>=0.9.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]
